making some changes in the login logic
